requirements.txt
pandas
psycopg2
sqlalchemy
python-dotenv
matplotlib
numpy
pickle5
datetime
boto3
scikit-learn